backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
openai==1.3.7
python-dotenv==1.0.0
httpx==0.27.0
sqlalchemy>=2.0.0
celery>=5.3.0
redis>=4.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# YouTube ingestion dependencies
youtube-transcript-api>=0.6.2
webvtt-py>=0.5.1
yt-dlp>=2024.08.06
